+++ src_core_extensions/lib.rs
@@ -117,7 +117,7 @@
 //! and [`quasiconst`] macro (for declaring types that emulate generic constants).
 //! Enables the `"generics_parsing"` feature.
 //!
-//! - `"macro_utils`: Enables the [`rewrap_opaque`] macro.
+//! - `"macro_utils`: Enables the [`rewrap_macro_parameters`] macro.
 //!
 //! - `"generics_parsing"`: 
 //! Enables the [`parse_generics`], [`parse_generics_and_where`],
@@ -214,7 +214,7 @@
 //! [`type_asserts`]: ./type_asserts/index.html
 //! [`type_level_bool`]: ./type_level_bool/index.html
 //!
-//! [`rewrap_opaque`]: ./macro.rewrap_opaque.html
+//! [`rewrap_macro_parameters`]: ./macro.rewrap_macro_parameters.html
 //! [`parse_generics`]: ./macro.parse_generics.html
 //! [`parse_generics_and_where`]: ./macro.parse_generics_and_where.html
 //! [`split_generics_and_where`]: ./macro.split_generics_and_where.html
@@ -472,7 +472,7 @@ pub mod __ {
     pub use std_::compile_error;
     
     #[cfg(feature = "macro_utils")]
-    pub use core_extensions_proc_macros::__priv_rewrap_opaque;
+    pub use core_extensions_proc_macros::__priv_rewrap_macro_parameters;
     
     #[cfg(feature = "enable_proc_macro_crate")]
     pub use core_extensions_proc_macros::{__priv_unwrap_bound, __priv_split_generics};