+++ src_core_extensions/macros/macro_utils.rs
@@ -1,16 +1,16 @@
-/// Rewraps opaque macro parameters into parentheses.
+/// Rewraps the tokens inside macro parameters into parentheses.
 /// 
 /// # Syntax
 /// 
-/// This macro transforms `~` immediately followed by an opaque macro parameter 
-/// (eg: a `$parameter:expr` parameter) into its tokens wrapped in parentheses.
+/// This macro transforms `~` immediately followed by a macro parameter 
+/// into its tokens wrapped in parentheses.
 /// 
 /// You can escape `~` by writing it twice (`~~`), returning a single `~` from the macro.
 /// 
 /// # Example
 /// 
 /// ```rust
-/// pub use core_extensions::rewrap_opaque;
+/// pub use core_extensions::rewrap_macro_parameters;
 /// 
 /// crate::constify!{
 ///     pub fn foo() -> u32 {
@@ -30,12 +30,12 @@
 /// #[macro_export]
 /// macro_rules! constify {
 ///     ($($item:item)*) => {
-///         rewrap_opaque!{
+///         rewrap_macro_parameters!{
 ///             $crate::__priv_constify_inner!{
 ///                 hello world ~~~~
 ///                 // `__priv_constify_inner` can't destructure `$item`,
-///                 // so you need to use `rewrap_opaque` and prefix the parameter with
-///                 // `~` to rewrap its tokensin parentheses
+///                 // so you need to use `rewrap_macro_parameters` and prefix the parameter with
+///                 // `~` to rewrap its tokens in parentheses
 ///                 $(~$item)*
 ///             }
 ///         }
@@ -45,7 +45,7 @@
 /// #[macro_export]
 /// #[doc(hidden)]
 /// macro_rules! __priv_constify_inner{
-///     (   
+///     (
 ///         hello world ~ ~
 ///         $((
 ///             $(#[$attr:meta])*
@@ -68,11 +68,11 @@
 /// ```
 /// 
 #[macro_export]
-macro_rules! rewrap_opaque {
+macro_rules! rewrap_macro_parameters {
     (
         $($tokens:tt)*
     ) => {
-        $crate::__::__priv_rewrap_opaque!{
+        $crate::__::__priv_rewrap_macro_parameters!{
             $($tokens)*
         }
     };