+++ tests/misc_tests_modules.rs
@@ -19,6 +19,9 @@ mod misc_tests {
     #[cfg(feature = "item_parsing")]
     mod item_parsing_tests;
 
+    #[cfg(feature = "macro_utils")]
+    mod macro_utils_tests;
+
     #[cfg(feature = "transparent_newtype")]
     mod transparent_newtype_tests;