+++ Changelog.md
@@ -4,6 +4,10 @@ This started being written for the 0.1.19 release, it doesn't cover versions pri
 
 # 1.0
 
+### 1.3.0
+
+Added the `count_tts` and `gen_idents` macros, under the `"macro_utils"` feature.
+
 ### 1.2.0
 
 Added `"macro_utils"` feature.